refactor: router 분리 작업 완료
